Screening Loop now iterates for file in 1:length.
Fixed fread error in screening R script loop caused by a temporary TSV file in directory.

In screening script, added linear residual function, (as well as a possible fix) as well as adding subjects to the dataframe of residuals.

Added freesurfer data of first 5 subjects, titled examplefreesurfoutput.

volumes_dt_screening was copied to a subset of this called subset_data.